refactor(starknet_integration_tests): consolidate config utils into executable setup
commit-id:21bf81ea
